requirements.txt
maze_generator_jotbleach
Pillow
numpy
torch
torchvision
accelerate>=0.34.0
transformers>=4.52.0
peft>=0.12.0